GPT-61 - Active Layers - Cleanup to use same store as KnownLayers
I realised that I removed the need to have a separate Store for the ActiveLayersPanel.  So I've dropped it and use the original layerStore
that the KnownLayersPanel uses.  This has the added benefit that the two way binding works properly now so when the layer is deleted
in either panel it is deleted from the other.  Noice!